libspdm: Build end_responder into no_std binary
Bundle the end_responder.key into the binary at build time for no_std
builds. This allows us to include the key in the binary for using it
with libspdm without requiring file system support.

This has the negative impact of increaing the binary and hard coding the
key at build time, but it at least allows us to support the same libspdm
functions with no_std.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>