request: handle responder unsupported capability
Log an error if the responder returns an unsupported capability. This
allows us to still setup a session and try sending a request.

For instance, if the responder is using SPDM version 1.0, using keyexhcange is
not supported in versions less than 1.1. We can log an error for such a
case here as it isn't critical, but still send another request
(get-version) later that isn't impaceted by this missing functionality.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>