mctp: add i2c transport
This adds i2c transport layer support for the tock-responder. It has
been tested on an nrf52840 based device (Particle Boron).

For spdm transaction to work, currently, as we do not yet support
chunking messages, we require the kernel I2C buffer to be sufficiently
large. That is, it should be >= `LIBSPDM_MAX_SPDM_MSG_SIZE`

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>