usb_i2c/mctp: use the maximum LIBSPDM_SPDM_MSG_SIZE
Since we use chunking, we can afford to use the larger buffers to
capture large transfers and limit the actual transfer to smaller frames.

For memory restrained targets, this should be reverted, but note that
for functionality such as `set_cert` larger buffers are required.

The primary motivation for this change is to support `set_cert` request.
Without this change libspdm will return an error due to insufficient
buffer size. This could possibly be addressed.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>