spdm: add limited support for 1.3 GEN_CSR
This implements the `libspdm_gen_csr_ex()` fn to allow us
to continue to support a `GET_CSR` in libspdm 1.3. This is limited
in the sense that we don't support CsrTrackingTags yet.

That is, a responder will immediately process the request and return the
CSR without needing a reset, as such, a requester must not request to
seek a CsrTrackingTag. Assertions are in place to catch such a case.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>