manifest: handle decode failures
Instead of panic-ing, let's return an error on cases where we are unable
to decode the measurement manifest.

This is useful, if the target platform does not have access to the decode
utilities (`cbor-diag`). Such targets can still use spdm-utils to extract
the raw-bitstream of the measurement manifest received, and skip
decoding all together.

Such a case should only occur when running in `requester` mode, for
examle and embedded platform with limited access to required libraries.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>