Add dynamic image measurement
Add a command line argument for responders to enable dynamic image
measurement. This allows us to more accuratly model a real respnoder.

At startup we spawn two threads to hash the "kernel" and the
application. We currently join them before running the responder, we
actually only need to do this before generating the measurement, but
there isn't a nice way to do that from the libspdm Rust library. I
thought of a closure callback, but it's all a bit messy so the current
approach will do for now.

`libspdm_fill_measurement_image_hash_block()` will now dynamically use a
type of image block based on the init status of the `Lazy`
`DYN_IMAGE_MEASURE`.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>
Signed-off-by: Alistair Francis <alistair.francis@wdc.com>