libspdm: spdm: Don't double count measurements
The measurement helper functions (like
libspdm_fill_measurement_image_hash_block()) already include the
spdm_measurement_block_dmtf_t size and the length of the data in the
return value. There is no need to calculate it twice.

This commit removes the second attempt to calculate total_size_needed.
This fixes the wrong size being reported if changes are made to the
measurement functions.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>