libspdm: Check the libspdm_write_certificate_to_nvm arguments
The libspdm version of libspdm_write_certificate_to_nvm() has changed,
the C header is different as is the bindgen Rust code.

The problem is, we don't get an error that our function no longer
matches. For now this is ok, as libspdm only appended arguments, but in
the future this could cause breakages.

There doesn't appear to be any reasonable way to assert that our
function matches the bindgen version. Moving
libspdm_write_certificate_to_nvm() to libspdm_rs.rs for example causes
errors as there are two copies, but that doesn't help us.

The best solution I can think of is to use a test function to check that
the two functions at least have the same function pointer. This is the
equivilent checks that a C compiler would do using function pointers.

This adds the checks for libspdm_write_certificate_to_nvm(), we also
need to do the check for all other functions next.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>