third-party: libspdm & validator bump to 3.4.0
Bumps `libspdm` and `SPDM-Responder-Validator` to the latest 3.4.0
release. Also updates the README(s) to document that we do not yet
support `LIBSPDM_ENABLE_CAPABILITY_MEL_CAP`. This capability is an
SPDM 1.3 feature, where a measurement extention log can be collected. As
such, it requires this functionality to be implemented in `spdm-utils`.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>