libspdm: Check all extern function arguments
We don't generate errors or warnings if the function signatures we implement
don't match the version in libspdm. So far this has been ok as libspdm only
appended arguments, but in the future this could cause breakages.

There doesn't appear to be any reasonable way to assert that our
function matches the bindgen version.

The best solution I can think of is to use a test function to check that
the two functions at least have the same function pointer. This is the
equivilent checks that a C compiler would do using function pointers.

This patch adds a macro that can generate that function (somewhat manually)
and then uses that macro for all of the `extern "C"` functions.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>