libspdm: Support spdm_context in the HAL functions
libspdm PR https://github.com/DMTF/libspdm/pull/2592 adds the
LIBSPDM_HAL_PASS_SPDM_CONTEXT macro which allows passing the
spdm_context to the HAL functions.

According to [1] this will eventually be the default, so let's update
now to be ready and then we can use the context in the HAL functions.

This updates the build instructions and CI to expose the spdm_context to
our `extern "C"` HAL functions.

1: https://github.com/bluerise/libspdm/blob/7dd35de89943090dfcabe39192af31fa26259a36/include/library/spdm_lib_config.h#L404

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>