requests: support sending user specified requests
This allows the user to specify a list of SPDM requests and `spdm-utils`
shall issue them in the order specified. This should be a useful tool
for CI, debugging and development.

It is the responsibility of the user to ensure the SPDM requests are
issued in a specification compliant way.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>