spdm_transport: generic: refactor IO buffers
This allows us to remove alot of repeated code for buffer `set/get`. This
approach uses a heap allocated vector and just passes a reference to
`libspdm` when required.

The associated memory shall be explicitly freed
by taking ownership when the buffers are no longer required. Which can be
done by calling `libspdm_drop_io_buffers()`.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>