manifest: check in serialised manifest
Responder platforms that do not support the generation of the
`manifest.out.cbor` file at build time using the ruby `cbor-diag` gem,
can use this as a pre-generated file that is checked in.

For example, responders that use buildroot, where there is no feasible
way to add support for `cbor-diag`.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>