doe_pci_cfg: add doe auto suggest support
If a device id and vendor id is not specified by user, scan the PCIe
bus for devices that support DoE capability. Then uggest a list
of options for the user to choose from. `spdm-utils` will read the
specified device index from stdin then fetch the corresponding device
and vendor ids.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>