Prepare for Rust 2024 edition
This preapres for the update to the Rust 2024 edition.

This is a fairly mechanical change, mostly just wrapping unsafe function
calls with unsafe instead of relying on the calling function being
unsafe.

This brings the number of unsafe keywords in our code base significantly
higher, which is a shame. Overall the actual unsafe code is the same,
just more accurately described.

Hopefully this can be improved in the future, especially if we can remove
some of the C style pointer operations.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>