requests: SET_CERTIFICATE: support secured session
For slots 1-7, if the provisioning happens in a trusted environment, the
Requester should issue SET_CERTIFICATE inside a secure session [1].

Currently, when a SET_CERTIFICATE request is invoked, `spdm-utils` already establishes
a session. So just pass the session ID in such that we conform to the spec.
We also enforce a check now to assert that slots 1-7 will only be
settable in a secured session.

[1] SPDM 1.3: Margin 769

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>