spdm: avoid unwrap when decoding cbor manifest
The process_cbor_measurement_manifest() is designed to be called
speculatively. In that, it is expected to return an error if the
manifest fails to decode (non or bad cbor). Currently, the unwraps will
cause valid, but non cbor measurements to fail. This patch removes the
unwraps and returns the error up.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>